+++ internal/function/registry.go
@@ -17,4 +17,5 @@ var Functions = []sql.Function{
 	sql.Function1{Name: "uast_children", Fn: NewUASTChildren},
 	sql.Function1{Name: "uast_imports", Fn: NewUASTImports},
 	sql.Function1{Name: "is_vendor", Fn: NewIsVendor},
+	sql.Function2{Name: "blame", Fn: NewBlame},
 }