+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added BLAME function.
+
 ## [0.24.0-rc2] - 2019-10-02
 
 ## Fixed